Calculates two levels of segregation: a multi-group segregation metric for each larger geographic unit supplied, as well as a decomposed version of the same metric for each smaller geographic unit. Only geographies that are perfectly nested within one another, e.g., tracts within counties, are supported. Note that all segregation calculations rely on `segregation` and users should refer to that package at https://github.com/elbersb/segregation for further implementation details.

Calculates two levels of segregation: a multi-group segregation metric for each larger geographic unit supplied, as well as a decomposed version of the same metric for each smaller geographic unit. Only geographies that are perfectly nested within one another, e.g., tracts within counties, are supported. Note that all segregation calculations rely on `segregation` and users should refer to that package at https://github.com/elbersb/segregation for further implementation details.

Usage

calculate_segregation_metrics(data, nesting_geography_geoid_length)

Arguments

data

A dataframe containing a `GEOID` column and the required input measures, e.g., of race or income, at a single geography (e.g., tract), formatted wide. the GEOID column must be a character column, and each GEOID must be unique. inputted data cannot contain other measures. for example: tibble::tribble( ~GEOID, ~race_nonhispanic_white_alone, ~race_nonhispanic_black_alone, "37001020100", 2835, 1035, "37001020200", 1205, 1321)

nesting_geography_geoid_length

The length of the GEOID that identifies nesting geographies. For example, if smaller_geography_data is defined at the tract level (with a GEOID of length 11), then nesting_geography_geoid_length = 5 would return segregation metrics for counties (which have a GEOID of length 5) and for tracts (relative to other tracts within the same county).

Value

A dataframe comprising segregation estimates and associated p-values at both geographic levels. Segregation is measured using the Mutual Information Index (M) at the larger geography and a decomposed version thereof at the smaller geography level.

See also

Examples

variables = c(
  race_nonhispanic_white_alone_ = "B03002_003",
  race_nonhispanic_black_alone_ = "B03002_004",
  race_nonhispanic_native_alone_ = "B03002_005",
  race_nonhispanic_asian_alone_ = "B03002_006",
  race_nonhispanic_nhpi_alone_ = "B03002_007")
test_data = tidycensus::get_acs(
  geography = "tract",
  state = "CT",
  variables = variables,
  output = "wide") %>%
  ## can only include a dataframe with a GEOID column and segregation-related measures
   dplyr::select(-NAME)
#> Getting data from the 2017-2021 5-year ACS
#> Warning: • You have not set a Census API key. Users without a key are limited to 500
#> queries per day and may experience performance limitations.
#> ℹ For best results, get a Census API key at
#> http://api.census.gov/data/key_signup.html and then supply the key to the
#> `census_api_key()` function to use it throughout your tidycensus session.
#> This warning is displayed once per session.
calculate_segregation_metrics(data = test_data, nesting_geography_geoid_length = 5)
#> Joining with `by = join_by(GEOID)`
#> # A tibble: 883 × 5
#>    GEOID       segregation_small_geog…¹ p_small_geography segregation_large_ge…²
#>    <chr>                          <dbl>             <dbl>                  <dbl>
#>  1 09001010101                   0.0940           0.00541                  0.150
#>  2 09001010102                   0.142            0.00501                  0.150
#>  3 09001010201                   0.155            0.00400                  0.150
#>  4 09001010202                   0.152            0.00584                  0.150
#>  5 09001010300                   0.0918           0.00495                  0.150
#>  6 09001010400                   0.0599           0.00702                  0.150
#>  7 09001010500                   0.0726           0.00640                  0.150
#>  8 09001010600                   0.0443           0.00233                  0.150
#>  9 09001010700                   0.0382           0.00389                  0.150
#> 10 09001010800                   0.136            0.00368                  0.150
#> # ℹ 873 more rows
#> # ℹ abbreviated names: ¹​segregation_small_geography,
#> #   ²​segregation_large_geography
#> # ℹ 1 more variable: p_large_geography <dbl>