Calculate constructs frequently used in social sciences research while leveraging [tidycensus::get_acs()] to acquire raw estimates from the Census Bureau API.

Calculate constructs frequently used in social sciences research while leveraging [tidycensus::get_acs()] to acquire raw estimates from the Census Bureau API.

Usage

compile_acs_data(
  variables = NULL,
  years = c(2022),
  geography = "county",
  states = NULL,
  counties = NULL,
  retain_moes = FALSE
)

Arguments

variables

A named vector of ACS variables such as that returned from [urbnindicators::list_acs_variables()].

years

A character vector (or coercible to the same) comprising one or more four-digit years for which to pull five-year American Community Survey estimates.

geography

A geography type that is accepted by [tidycensus::get_acs()], e.g., "tract", "county", "state", among others.

states

A vector of one or more state names, abbreviations, or codes as accepted by [tidycensus::get_acs()].

counties

A vector of one or more county names, abbreviations, or codes as accepted by [tidycensus::get_acs()]. NOTE: this parameter is not currently supported and must be set to NULL (as it is by default).

retain_moes

Boolean. Include margins of error (MOE) in the returned dataframe, or omit them?

Value

A dataframe containing the requested `variables`, their MOEs (optionally), a series of derived variables, such as percentages, and the year of the data. Returned data are formatted wide.

See also

Examples

acs_variables = list_acs_variables(year = "2022")
compile_acs_data(
  variables = acs_variables,
  years = c(2021, 2022),
  geography = "county",
  states = "NJ",
  counties = NULL,
  retain_moes = TRUE)
#> Warning: Variable names and geographies for ACS data products can change between years.
#> Changes to geographies are particularly significant across decades (e.g., from 2019 to 2020), but these changes can occur in any year.
#> Users should ensure that the logic embedded in this function--which was developed around five-year ACS estimates for 2017-2021--remains accurate for their use cases.
#> Evaluation of measures and geographies over time should be thoroughly quality checked.
#> Getting data from the 2017-2021 5-year ACS
#> Getting data from the 2018-2022 5-year ACS
#> # A tibble: 42 × 1,326
#>    GEOID NAME                      total_population_uni…¹ public_assistance_un…²
#>    <chr> <chr>                                      <dbl>                  <dbl>
#>  1 34001 Atlantic County, New Jer…                 273865                 105277
#>  2 34003 Bergen County, New Jersey                 952979                 348674
#>  3 34005 Burlington County, New J…                 460102                 172400
#>  4 34007 Camden County, New Jersey                 521194                 196939
#>  5 34009 Cape May County, New Jer…                  95488                  41971
#>  6 34011 Cumberland County, New J…                 154921                  51873
#>  7 34013 Essex County, New Jersey                  852720                 307613
#>  8 34015 Gloucester County, New J…                 300821                 109290
#>  9 34017 Hudson County, New Jersey                 713264                 282832
#> 10 34019 Hunterdon County, New Je…                 128807                  48975
#> # ℹ 32 more rows
#> # ℹ abbreviated names: ¹​total_population_universe, ²​public_assistance_universe
#> # ℹ 1,322 more variables: public_assistance_received <dbl>,
#> #   snap_universe <dbl>, snap_received <dbl>,
#> #   household_income_quintile_upper_limit_1 <dbl>,
#> #   household_income_quintile_upper_limit_2 <dbl>,
#> #   household_income_quintile_upper_limit_3 <dbl>, …